fix: split auth helpers so server.ts doesn't pull in next/headers
server.ts is a custom Node entrypoint loaded via tsx, not Next.js. When
it imported '@/lib/auth' to get assertAuthSecret() and
getSessionFromCookieHeader(), it transitively pulled
\`import { cookies } from 'next/headers'\`, which loads
next/dist/server/app-render/async-local-storage.js. Under tsx the
AsyncLocalStorage instance isn't initialized at that point and Next
throws "Invariant: AsyncLocalStorage accessed in runtime where it is
not available", killing \`npm run dev\` before it can start.

Split src/lib/auth.ts into:
- src/lib/auth/session.ts — runtime-agnostic JWT/cookie helpers. No
  next/* imports. Safe to load from server.ts, middleware.ts, MCP, and
  Next route handlers alike.
- src/lib/auth.ts — thin shim that adds login() (the only function that
  needs cookies() from next/headers) and re-exports decrypt /
  getSessionFromCookieHeader / readSessionCookie for existing callers.

server.ts and middleware.ts now import from '@/lib/auth/session'
directly. Existing imports from '@/lib/auth' (login route, oidc
callback, discovery merge route, tests) continue to work via the shim.

Verified: \`npx tsx server.ts\` reaches "> Ready on http://localhost:3000"
on Node 20.20.2. Test count unchanged (189 pass, 3 skipped).
